@@ -3,8 +3,8 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "react": "^16.6.3",
-    "react-dom": "^16.6.3",
+    "react": "^16.11.0",
+    "react-dom": "^16.11.0",
     "react-scripts": "2.1.1",
     "react-textarea-autosize": "^7.1.2",
     "semantic-ui-css": "^2.4.1",
