@@ -1,4 +1,4 @@
-import React, { Component } from "react";
+import React, { useState, useEffect } from "react";
 import {
   Button,
   Divider,
@@ -8,15 +8,16 @@ import {
   Segment
 } from "semantic-ui-react";
 import TextareaAutosize from "react-textarea-autosize";
-import SimpleStorageContract from "./contracts/SimpleStorage.json";
+import TLDRContract from "./contracts/TLDR.json";
 import getWeb3 from "./utils/getWeb3";
-
 import "./App.css";
 
-class App extends Component {
-  state = { storageValue: 0, web3: null, accounts: null, contract: null };
+export default function App() {
+  const [web3, setWeb3] = useState(0);
+  const [accounts, setAccounts] = useState();
+  const [contract, setContract] = useState();
 
-  componentDidMount = async () => {
+  const fetchData = async () => {
     try {
       // Get network provider and web3 instance.
       const web3 = await getWeb3();
@@ -26,27 +27,26 @@ class App extends Component {
 
       // Get the contract instance.
       const networkId = await web3.eth.net.getId();
-      const deployedNetwork = SimpleStorageContract.networks[networkId];
+      const deployedNetwork = TLDRContract.networks[networkId];
       const instance = new web3.eth.Contract(
-        SimpleStorageContract.abi,
+        TLDRContract.abi,
         deployedNetwork && deployedNetwork.address
       );
-
-      // Set web3, accounts, and contract to the state, and then proceed with an
-      // example of interacting with the contract's methods.
-      this.setState({ web3, accounts, contract: instance }, this.runExample);
+      setWeb3(web3)
+      setAccounts(accounts);
+      setContract(instance);
     } catch (error) {
-      // Catch any errors for any of the above operations.
       alert(
         `Failed to load web3, accounts, or contract. Check console for details.`
       );
       console.error(error);
     }
-  };
+  }
+  useEffect(() => {
+    fetchData()
+  });
 
-  runExample = async () => {
-    const { accounts, contract } = this.state;
-
+  const submitLS = async () => {
     // Stores a given value, 5 by default.
     await contract.methods.set(5).send({ from: accounts[0] });
 
@@ -57,40 +57,38 @@ class App extends Component {
     this.setState({ storageValue: response });
   };
 
-  render() {
-    if (!this.state.web3) {
-      return <div>Loading Web3, accounts, and contract...</div>;
-    }
-    return (
-      <div className="App">
-        <Container>
-          <Divider hidden />
-          <Header as="h1">🖋️ TLDR</Header>
-          <Header as="h3">The lexDAO Registry</Header>
-
-          <Segment>
-            <Form>
-              <Form.Field>
-                <label>Lex Address (who do the payments go to?)</label>
-                <input placeholder="0x0000000000000000000000000000000000000000" />
-              </Form.Field>
-              <Form.Field>
-                <label>Lex Rate (1 = 0.01%)</label>
-                <input placeholder="0" />
-              </Form.Field>
-              <Form.Field
-                control={TextareaAutosize}
-                label="Template Terms"
-                placeholder="Your Template Here..."
-                useCacheForDOMMeasurements
-              />
-              <Button type="submit">Submit</Button>
-            </Form>
-          </Segment>
-        </Container>
-      </div>
-    );
+  if (!web3) {
+    return <div>Loading Web3, accounts, and contract...</div>;
   }
-}
+  return (
+    <div className="App">
+      <Container>
+        <Divider hidden />
+        <Header size="huge" as="h1">
+          🖋️ TLDR
+        </Header>
+        <Header as="h3">The lexDAO Registry</Header>
 
-export default App;
+        <Segment>
+          <Form>
+            <Form.Field>
+              <label>Lex Address (who do the payments go to?)</label>
+              <input placeholder="0x0000000000000000000000000000000000000000" />
+            </Form.Field>
+            <Form.Field>
+              <label>Lex Rate (1 = 0.01%)</label>
+              <input placeholder="0" />
+            </Form.Field>
+            <Form.Field
+              control={TextareaAutosize}
+              label="Template Terms"
+              placeholder="Your Template Here..."
+              useCacheForDOMMeasurements
+            />
+            <Button type="submit">Submit</Button>
+          </Form>
+        </Segment>
+      </Container>
+    </div>
+  );
+}
